Clear the Linux-only clippy warnings that workspace linting surfaces
All four are pre-existing on main and sit in code the previous
root-package clippy scope never reached — three of them behind
cfg(target_os = "linux"), so they are invisible on a macOS host.

- main.rs: drop a duplicate allow(too_many_arguments) on
  run_in_keepalive_container; the function already carries one next to
  its signature, and the stray copy sat above the doc comment
- pty.rs: collapse the nested dup2 guard into one condition
- storage.rs: drop two no-op `as u64` casts on statvfs arithmetic

+++ crates/smolvm-agent/src/storage.rs
@@ -3873,7 +3873,7 @@ fn get_disk_usage(path: &Path) -> Result<(u64, u64)> {
             let free = stat.f_bfree * stat.f_frsize;
             let used = total - free;
 
-            Ok((total as u64, used as u64))
+            Ok((total, used))
         }
     }
